device: Fix infinite loop when pairing
When pairing with some devices (such as GSConnect), mconnect would get stuck in
a loop of sending and receiving pair packets. This would start when we sent a
pair packet on new connections, even if we were already paired with the other
device.

Fixed by not sending a pair packet on new connections if we're already paired.